Extract payment method order note tests into a separate test suite
